requirements.txt
absl-py==2.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
authlib==1.6.6
av==16.0.1
beartype==0.22.9
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.3
ctranslate2==4.6.1
cycler==0.12.1
cyclopts==4.4.4
diskcache==5.6.3
dnspython==2.8.0
docstring-parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
exceptiongroup==1.3.1
fakeredis==2.33.0
fastapi==0.128.0
faster-whisper==1.2.1
fastmcp==2.14.2
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.1.4
humanfriendly==10.0
idna==3.11
importlib-metadata==8.7.1
jaraco-classes==3.4.0
jaraco-context==6.0.2
jaraco-functools==4.4.0
jax==0.8.0
jaxlib==0.8.0
jsonschema==4.26.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.7.0
kiwisolver==1.4.9
lupa==2.6
markdown-it-py==4.0.0
matplotlib==3.10.7
mcp==1.25.0
mdurl==0.1.2
mediapipe==0.10.14
ml-dtypes==0.5.3
more-itertools==10.8.0
mpmath==1.3.0
numpy==2.3.4
onnxruntime==1.23.2
openapi-pydantic==0.5.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opentelemetry-api==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-instrumentation==0.60b1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opt-einsum==3.4.0
packaging==25.0
pathable==0.4.4
pathvalidate==3.3.1
pillow==12.0.0
piper-tts==1.3.0
platformdirs==4.5.1
prometheus-client==0.23.1
protobuf==4.25.8
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pydocket==0.16.3
pygments==2.19.2
pyjwt==2.10.1
pyparsing==3.2.5
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.21
pyyaml==6.0.3
redis==7.1.0
referencing==0.36.2
requests==2.32.5
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.30.0
scipy==1.16.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.3
sse-starlette==3.1.2
starlette==0.50.0
sympy==1.14.0
tokenizers==0.22.1
tqdm==4.67.1
typer==0.21.1
typer-slim==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.3.0
uvicorn==0.40.0
websockets==15.0.1
wrapt==1.17.3
zipp==3.23.0
